ad fixes
